requirements.txt
kivy==2.3.0
pyobjc
